pyproject.toml
[project]
name = "issue-condenser"
version = "0.1.0"
description = ""
authors = [
    {name = "alireza",email = "alirezzzhp1378@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygithub (>=2.7.0,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "tree-sitter (>=0.25.1,<0.26.0)",
    "tree-sitter-go (>=0.23.4,<0.24.0)",
    "openai (>=1.99.1,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
